expense_tracker/requirements.txt
Django>=4.2,<5.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
django-environ==0.11.2
whitenoise==6.7.0
# optional for Prometheus metrics later
django-prometheus==2.3.1
